Adding Controller
A simple Changes that describe How to connect an Ionic Angular Project to the Mysql Database and how to use PHP APIs also demonstrate how the MVC work ionic and angular. used some controller and plugins to add barcode generator and Barcode Reader to scan and save the data to local storage.